requirements.txt
newspaper3k==0.2.8
sendgrid==6.10.0
jinja2==3.1.2
beautifulsoup4==4.12.2
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
google-cloud-firestore==2.13.1
google-auth==2.23.4
pandas==2.1.1
aiohttp==3.9.1
python-dateutil==2.8.2
llama-index
llama-index-readers-web